Move sympy.Functions to _sympy/functions.py
[ghstack-poisoned]
